# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cordy API",
    "version": "1.0.0",
    "description": "API documentation for Cordy"
  },
  "servers": [
    {
      "url": "https://coordy-prod.vercel.app/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000/api",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste your JWT token here (format: Bearer <token>)"
      }
    },
    "schemas": {
...
      "Event": {
        "type": "object",
        "required": [
          "title",
          "date"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique ID of the event"
          },
          "title": {
            "type": "string",
            "example": "Team Sync"
          },
          "notes": {
            "type": "string",
            "example": "Weekly stand-up meeting"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-21T09:00:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "WAITING_RESPONSE",
              "NEED_ACTION",
              "COMPLETED"
            ],
            "example": "WAITING_RESPONSE"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "example": "LOW"
          },
          "timezone": {
            "type": "string",
            "example": "WIB"
          },
          "availableTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "example": [
              "2025-10-22T09:00:00Z",
              "2025-10-23T10:00:00Z"
            ]
          },
          "matchedTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          },
          "selectedTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john@example.com"
                }
              }
            }
          }
        }
      },
...
      "VoteRequest": {
        "type": "object",
        "required": [
          "selectedTimes"
        ],
        "properties": {
          "selectedTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "description": "List of time slots chosen by the participant"
          }
        },
        "example": {
          "selectedTimes": [
            "2025-10-22T09:00:00.000Z",
            "2025-10-22T11:00:00.000Z"
          ]
        }
      },
      "VoteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Vote submitted successfully"
          },
          "matchedTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Times matched across all participants"
          },
          "updatedParticipant": {
            "type": "object",
            "description": "Updated participant details after submission"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/event": {
      "post": {
        "summary": "Create a new event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/event": {
      "get": {
        "summary": "Get all events for the authenticated user",
        "description": "Retrieve all events created by the authenticated user. Supports filtering by title, status, and date (year, month, day).",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "Filter by event title (case-insensitive)"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "WAITING_RESPONSE",
                "NEED_ACTION",
                "COMPLETED"
              ]
            },
            "description": "Filter by event status"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by year (e.g. 2025)"
          },
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by month (1–12)"
          },
          {
            "in": "query",
            "name": "day",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by specific day of the month"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "The number of items to skip before starting to collect the result set"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "The number of items to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved events"
          },
          "400": {
            "description": "Invalid query parameter"
          }
        }
      }
    },
    "/event/{id}": {
      "get": {
        "summary": "Get a single event by ID",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an event by ID",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "patch": {
        "summary": "Update event details",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/event/slug/{slug}": {
      "get": {
        "summary": "Get an event by its slug (Public)",
        "tags": [
          "Events"
        ],
        "description": "Public endpoint — no authentication required.  \nUsed by App Clip or shared links to access an event without logging in.\n",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Unique event slug (e.g., \"team-sync-meeting\")",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found successfully"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/event/{eventId}/pick-time": {
      "patch": {
        "summary": "Pick the final time for an event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "selectedTime": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-10-30T09:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event time selected and marked as COMPLETED"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
...
    "/vote": {
      "post": {
        "summary": "Submit participant votes for an event (Public)",
        "tags": [
          "Votes"
        ],
        "description": "Allows participants to submit selected time slots for an event using query parameters.\nThis endpoint does **not** require authentication.\n",
        "parameters": [
          {
            "name": "event",
            "in": "query",
            "required": true,
            "description": "Unique slug of the event",
            "schema": {
              "type": "string",
              "example": "weekly-team-meeting"
            }
          },
          {
            "name": "participant",
            "in": "query",
            "required": true,
            "description": "Unique slug for the participant",
            "schema": {
              "type": "string",
              "example": "john-doe"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vote submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or already submitted"
          },
          "404": {
            "description": "Event or participant not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/vote/{eventLetter}/{participantLetter}": {
      "post": {
        "summary": "Submit votes via short App Clip link (Public)",
        "tags": [
          "Votes"
        ],
        "description": "Allows participants to submit votes via compact short links such as `/vote/b/a`.\nThis endpoint also does **not** require authentication.\n",
        "parameters": [
          {
            "name": "eventLetter",
            "in": "path",
            "required": true,
            "description": "Short slug representing the event",
            "schema": {
              "type": "string",
              "example": "b"
            }
          },
          {
            "name": "participantLetter",
            "in": "path",
            "required": true,
            "description": "Short slug representing the participant",
            "schema": {
              "type": "string",
              "example": "a"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vote submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or already submitted"
          },
          "404": {
            "description": "Event or participant not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}